Enrichment dotplot for positive or negative GSEA results

Plot enrichment dotplot for positive or negative GSEA results.

Usage

GSEA_dotplot(
  enrich.obj,
  type = c("pos", "neg"),
  show.term.num = 15,
  Selct.P = c("FDR", "P"),
  cutoff.P = 0.05,
  colors = c("blue", "white", "red"),
  size.range = c(3, 8),
  y.label.position = "right",
  title = NULL,
  legend.position = "right",
  ggtheme = theme_bw(base_rect_size = 1),
  ...
)

Arguments

enrich.obj

A GSEA enrichment object from clusterProfiler.

type

Specify whether you want to show positive or negative results.

show.term.num

A number or a list of terms. If it is a number, the first n terms will be displayed. If it is a list of terms, the selected terms will be displayed.

Selct.P

Selct.P P value (P) or adjust P value (FDR) were selected to define significant terms.

cutoff.P

A cutoff value for Select.P.

colors

A color vector for the dots.

size.range

Two numeric variables, the first is minimal value and the first is maximal value.

y.label.position

Y label position. right or left.

title

Title of the plot.

legend.position

Position of legend. 'none', 'right', 'left' or two numeric variables.

ggtheme

ggtheme of plot.

...

Other argument of ggplot2::theme.

Examples

if (FALSE) { # \dontrun{
data(geneList, package="DOSE")
GSEA_res <- simple_GSEA(geneList, gene.type = "ENTREZID", enrich.type = 'GO')
GSEA_dotplot(res$simplyGO)
} # }