Over-representative analysis

Perform over-representative analysis included GO, KEGG, WikiPathways, Reactome, MsigDB, Disease Ontoloty, Cancer Gene Network, DisGeNET, CellMarker, and CMAP.

Usage

simple_ORA(
  genes,
  background.genes = NULL,
  gene.type = "SYMBOL",
  enrich.type = c("GO", "KEGG", "MKEGG", "WikiPathways", "Reactome", "MsigDB", "DO",
    "CGN", "DisGeNET", "CellMarker", "CMAP"),
  organism = c("Human", "Mouse"),
  GO.ont = c("BP", "CC", "MF", "ALL"),
  GO.simplify = T,
  KEGG.use.internal.data = F,
  MsigDB.category = "H",
  CMAP.min.Geneset.Size = 3,
  pvalue.cutoff = 0.05,
  qvalue.cutoff = 0.05,
  padjust.method = "BH",
  min.Geneset.Size = 10,
  max.Geneset.Size = 1000
)

Arguments

genes

A vector of gene id.

background.genes

Background genes. If missing, the all genes listed in the database (eg TERM2GENE table) will be used as background.

gene.type

gene id type input gene.

enrich.type

Select an enrichment method. One of GO, KEGG, MKEGG, WikiPathways, Reactome, MsigDB, DO, CGN, DisGeNET, CellMarker, and CMAP.

organism

Specify species, currently support only Human and Mouse.

GO.ont

GO parameter. One of "BP", "MF", and "CC" subontologies, or "ALL" for all three.

GO.simplify

GO parameter. Whether to remove redundancy of enriched GO terms. If TRUE, will return a list with raw result and result after remove redundancy.

KEGG.use.internal.data

KEGG parameter. Logical, use KEGG.db or latest online KEGG data.

MsigDB.category

MsigDB parameter. MSigDB collection abbreviation, such as All, H, C1, C2, C3, C4, C5, C6, C7.

CMAP.min.Geneset.Size

CMAP parameter. Minimal size of CMAP genes annotated for testing. Recommended use 3.

pvalue.cutoff

pvalue cutoff on enrichment tests to report as significant.

qvalue.cutoff

qvalue cutoff on enrichment tests to report as significant.

padjust.method

one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none".

min.Geneset.Size

Minimal size of genes annotated for testing. Not suitable for CMAP.

max.Geneset.Size

Maximal size of genes annotated for testing.

Examples

genes <- c("CANX", "HSPA1B", "KLRC2", "PSMC6", "RFXAP", "TAP1")
res <- simple_ORA(genes, enrich.type = "GO")
#> ℹ Updating gene symbols...
#> Maps last updated on: Fri May 17 15:09:37 2024
#> ℹ Transforming "SYMBOL" to ENTREZID...
#> 'select()' returned 1:1 mapping between keys and columns
#> ℹ Performing GO-"BP" enrichment...
#> ℹ 6 significant terms were detected...
#> ℹ Symplifying GO results...
#> ℹ Return a list with raw enrich result and symplify result ...
#> ✔ Done!