Update gene symbol by HGNC

Update the gene symbol by HGNChelper

Usage

update_symbol(genes, species = c("Human", "Mouse"), unmapGene_keep = F)

Arguments

genes

a vector of gene symbol.

species

A character vector of length 1, either "Human" (default) or "Mouse"

unmapGene_keep

whether remove unmapped gene symbol.

Examples

human <- c("FN1", "TP53", "UNKNOWNGENE","7-Sep", "9/7", "1-Mar", "Oct4",
           "4-Oct", "OCT4-PG4", "C19ORF71", "C19orf71")
update_symbol(genes = human, unmapGene_keep = TRUE)
#> Maps last updated on: Fri May 17 15:09:37 2024
#> ! There have genes mapped to more than one suggested symbol, we only keep the first one ...
#>          ID   Symbol
#> 1       FN1      FN1
#> 2      TP53     TP53
#> 4     7-Sep  SEPTIN7
#> 5       9/7  SEPTIN7
#> 7      Oct4   POU5F1
#> 8     4-Oct   POU5F1
#> 9  OCT4-PG4 POU5F1P4
#> 10 C19ORF71  TEKTIP1
#> 11 C19orf71  TEKTIP1
#> 12    1-Mar  MARCHF1